Add support for replaying history to registered listeners
This commit updates the implementation to support the event.history service property for event handler services. Event history is replayed up to the number of requested historical events. Note that to efficiently handle this the event history buffer had to be re-ordered to be most recent event first rather than oldest event first

Signed-off-by: Tim Ward <timothyjward@apache.org>