Add support for Records as events
This commit makes the Aries Typed Event bundle a multi-release bundle which works on Java 8, but adds support for Record types on Java 16 or later. Event Publishers and Event Handlers can use Record types as normal.

The build now requires Java 17 (the LTS release) as a result of this change.

Signed-off-by: Tim Ward <timothyjward@apache.org>